pyproject.toml
[project]
name = "salesforce-mcp"
version = "0.1.0"
description = "salesforce-mcp runner"
authors = [{ name = "Salesforce MCP", email = "daniyalamir2003@gmail.com" }]
dependencies = [
    "mcp",
    "simple-salesforce",
    "python-dotenv",
    "requests",
    "streamlit",
]

[project.scripts]
salesforce = "server:main"